Edits on the HTML and CSS files
This will add the script tag for linking the JS file in the project.
In the CSS file, a height and width were defined for the container, and
this made the container be more responsive to changes on the screen
size. There was also changes for the margins in the image tag, and
changes to the font-size of the h1 tag.