Add more styles to thank you container
More styles added to the thank you container in the CSS file.
Mostly fitting the image properly in its container. Removed the image
tag for the illustration, and instead put the image in a div
and inserted it using the background-image property. This made an issue
with fitting the image in its container properly, much simpler.