Styling completed for project
Adjustments were made to the the margins for the ".rating-btn" buttons,
as well as the submit button. Margin adjustment for the ."thank-you-
image" image. Cursor pointer effects were added using CSS to the rating
and submit buttons. Margin-bottom was added to the h1 in the thank you
container. The text in the p tag of the thank you container was
centered, and font-size was added. ".seleted-rating" margin was
adjusted. Display flex was added to the ".attribution" class, and
centered horizontally, but currently unsure of what to do with this to
bring it down to the bottom.