More styles and starting JS file
On the JS file the containers and buttons potentially needed for the
active states and work that is going to be done, were grabbed
with query selectors. Ran into a issue where one can add the
"active" class to the button, however there is uncertainty on how to
remove it when a different button is clicked.