CSS and JavaScript event listeners added
Added :active pseudo classes in the CSS file on the ".rating-btn"
class in order to scale the button and give a popout effect when the
button is clicked. In the JavaScript file a event listener was added
to the "btnRatingContainer" element, in order to track and handle
which one of the buttons was clicked. As well adding or removing the
"active" class accordingly, depending on which button was clicked or not
clicked. Active pseudo class added to submit button for styling.